moved printBoard func to class Game
